Keep the user namespace from the host
Adding the "--userns keep-id" option and the "--group-add wheel" option to
our podman invocation allows the script run in the container to modify the
contents of the mounted in directories.

This in turn allows us to run the jenkins tests locally against the contents
of our current checkout as our regular user

This also conditionally mounts in the tmp directory used by jenkins in
it's workspaces. If that directory doesn't exist locally this script would
fail prior.